requirements.txt
# =============================================================================
# FastAPI Project Dependencies
# =============================================================================

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0

# Database
sqlmodel>=0.0.14
aiosqlite>=0.19.0
asyncpg>=0.29.0
aiomysql>=0.2.0

# Redis
redis>=5.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Utilities
python-multipart>=0.0.6
email-validator>=2.1.0

# Logging
python-json-logger>=2.0.0
loguru>=0.7.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Type checking
mypy>=1.8.0
ruff>=0.1.0
sqlmodel>=0.0.14
aiosqlite>=0.19.0
asyncpg>=0.29.0
aiomysql>=0.2.0

# Redis
redis>=5.0.0

# Utilities
python-multipart>=0.0.6
email-validator>=2.1.0

# Logging
python-json-logger>=2.0.0
loguru>=0.7.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
